# API.yaml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/SVGTOPPT/SVGToPPTAPI/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /configuration:
    get:
      tags:
      - conversion-configuration
      - router
      summary: Jelenlegi konfiguráció lekérése
      responses:
        "200":
          description: Sikeres lekérés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'
    put:
      tags:
      - conversion-configuration
      - router
      summary: Konfiguráció módosítása
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Configuration'
      responses:
        "200":
          description: Sikeres módosítás
...
  /files:
    get:
      tags:
      - conversion-configuration
      - router
      summary: Feltöltött fájlok listájának lekérdezése
      responses:
        "200":
          description: Sikeres fájl lista lekérdezés
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
    post:
      tags:
      - conversion-configuration
      - router
      summary: Új fájl feltöltése
      requestBody:
        content:
          application/vnd.ms-powerpoint:
            schema:
              $ref: '#/components/schemas/BinaryFile'
          application/vnd.openxmlformats-officedocument.presentationml.presentation:
            schema:
              $ref: '#/components/schemas/BinaryFile'
      responses:
        "200":
          description: Sikeres feltöltés
          content:
            schema:
              $ref: '#components/schemas/SingleId'
...
components:
  schemas:
...
    BinaryFile:
      type: string
      format: binary
    Configuration:
      type: object
      properties:
        afterSlide:
          type: integer
          description: Ez után a dia utánra lesznek beszúrva a képek
        maxImages:
          type: integer
          description: Legfeljebb ennyi kép szúrható be
        animation:
          type: string
          enum:
          - simple
        pptFileId:
          type: string
        svgFileId:
          type: string
          nullable: false